+++ jenkviz/crawl.py
@@ -8,14 +8,15 @@ __copyright__ = "Copyright (C) 2012 Nuxeo SA <http://nuxeo.com>"
 
    Crawl a Jenkins to extract builds flow.
 """
+import sys
 import os
 import re
 import logging
-from urlparse import urlparse
 from requests import get as http_get
 from model import Build
 from util import str2id
 from util import extract_token
+from util import split_jenkins_url
 from model import get_build
 from model import save_build
 from model import update_build
@@ -28,12 +29,17 @@ class Crawl(object):
         self.builds = {}
         self.count = 0
         self.root = None
+        if options.verbose:
+            self.req_config = {'verbose': sys.stderr}
+        else:
+            self.req_config = {'verbose': None}
 
     def crawl(self, url):
         self.url = url
-        u = urlparse(url)
-        self.path_root = u.path
-        self.server_url = url[:-len(self.path_root)]
+        path = split_jenkins_url(url)
+        self.server_url = path[0]
+        self.base_path = path[2]
+        self.path_root = path[1]
         self._crawl(self.path_root)
         self.clean(self.root)
         self.root.extra = self.stats()
@@ -58,9 +64,9 @@ class Crawl(object):
 
     def reverse_crawl(self, url):
         self.url = url
-        u = urlparse(url)
-        self.path_leaf = u.path
-        self.server_url = url[:-len(self.path_leaf)]
+        ret = split_jenkins_url(url)
+        self.server_url = ret[0]
+        self.path_leaf = ret[1]
         self.roots = []
         self._reverse_crawl(self.path_leaf)
         # TODO: disable cache and craw from roots
@@ -111,8 +117,13 @@ class Crawl(object):
         if self.options.from_file:
             body = self.fetch_build_from_file(url)
         else:
-            body = self.fetch_build_from_server(url)
+            body = self.fetch_from_server(url)
         ret = self.parse_build(url, body)
+        children = self.get_rest_children(ret)
+        for child in ret.get_downstream():
+            if child not in children:
+                children.append(child)
+        ret.set_downstream(children)
         if ret and self.options.to_file:
             self.save_build_to_file(ret, body)
         return ret
@@ -126,11 +137,13 @@ class Crawl(object):
     def update_build_to_db(self, build):
         update_build(self.db, build)
 
-    def fetch_build_from_server(self, url):
-        response = http_get(self.server_url + url)
-        if response.status_code != 200:
+    def fetch_from_server(self, url, params=None, ok_codes=None):
+        response = http_get(self.server_url + url, data=params, config=self.req_config)
+        if ok_codes is None:
+            ok_codes = [200, ]
+        if response.status_code not in ok_codes:
             logging.error('Failure: %s%s return %s' % (self.server_url, url, response.status_code))
-            return "ERROR: %s" % response.status_code
+            return "ERROR: %s: %s" % (response.status_code, response.text)
         return response.text
 
     def fetch_build_from_file(self, url):
@@ -146,10 +159,41 @@ class Crawl(object):
         file_path = os.path.join(dir_path, build_id + '.txt')
         open(file_path, 'w+').write(body.encode('utf-8'))
 
+    def get_rest_children(self, build):
+        """Use the REST API to discover children build."""
+        job_url = self.base_path + "/" + build.name
+        job_api_url = job_url + '/api/xml?xpath=//downstreamProject/name/text()'
+        jname_pattern = re.compile('^[0-9a-zA-Z\-\_]+$')
+        jbuild_pattern = re.compile('^[0-9]+$')
+        # 1. query the job page to get a list of downstream jobs
+        body = self.fetch_from_server(job_api_url, ok_codes=[200, 404, 500])
+        children_names = []
+        match = re.findall(r'matched ([0-9]*) ', body)
+        if len(match):
+            for i in range(1, int(match[0]) + 1):
+                body = self.fetch_from_server(job_url + '/api/xml?xpath=//downstreamProject[%d]/name/text()' % i,
+                                              ok_codes=[200, 404])
+                if jname_pattern.match(body.strip()):
+                    children_names.append(str(body))
+                    # print "  Downstream job: %s" % body
+        elif jname_pattern.match(body.strip()):
+            children_names.append(str(body))
+        ret = []
+        # 2. for each downstream ask a matching upstream build
+        for child in children_names:
+            url = self.base_path + "/" + child
+            url += '/api/xml?depth=1&xpath=//build[action/cause/upstreamProject="%s"%%20and%%20action/cause/upstreamBuild=%s]/number/text()' % (build.name, build.build_number)
+            body = self.fetch_from_server(url, ok_codes=[200, 404])
+            if jbuild_pattern.match(body.strip()):
+                ret.append(self.base_path + "/" + str(child) + "/" + str(body) + "/")
+                # print " Found downstream build: %s %s" % (child, body)
+        return ret
+
     def parse_build(self, url, body):
         if body.startswith('ERROR'):
-            name = url.split('/')[-3]
-            build_number = url.split('/')[-2]
+            ret = split_jenkins_url(url)
+            name = ret[3]
+            build_number = ret[4]
             return Build(url, body, name, build_number, None, None, 'Unknown', [],
                          self.server_url, '', [])
         name = extract_token(body, '<title>', ' ')