+++ CHANGES.rst
@@ -10,7 +10,32 @@ jenkviz git master
 
 :git: https://github.com/bdelbosc/jenkviz
 
-:Target: 0.2.1
+:Target: 0.3.1
+
+
+
+jenkviz 0.2.0
+------------------
+
+:Package: http://pypi.python.org/packages/source/j/jenkviz/jenkviz-0.3.0.tar.gz
+
+:github: https://github.com/bdelbosc/jenkviz/tree/0.3.0
+
+:Released date: 2012-01-31
+
+Upgrade notes
+~~~~~~~~~~~~~~
+
+* You should use the ``--update`` options or remove your existing
+  database ``~/jenkviz.db``.
+
+
+New Features
+~~~~~~~~~~~~~
+
+* Use the Jenkins REST API to discover downstream builds, this
+  make it work with free style job.
+ 
 
 Bug Fixes
 ~~~~~~~~~~